Added Model View Controller sub directories for organisation. Refactored code. All tests Passing. Updated UML class diagram
