+++ ConwaysGameOfLife/Controller/Orientations/Strategies/BottomLeftCornerStrategy.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
+using ConwaysGameOfLife.Model;
 
-namespace ConwaysGameOfLife.Orientations
+namespace ConwaysGameOfLife.Controller.Orientations.Strategies
 {
     public static class BottomLeftCornerStrategy
     {
