+++ ConwaysGameOfLife/Controller/Orientations/Strategies/BottomRightCornerStrategy.cs
@@ -1,14 +1,15 @@
 using System.Collections.Generic;
+using ConwaysGameOfLife.Model;
 
-namespace ConwaysGameOfLife.Orientations
+namespace ConwaysGameOfLife.Controller.Orientations.Strategies
 {
-    public static class BottomSideStrategy
+    public static class BottomRightCornerStrategy
     {
         public static List<State> GetNeighbourCellsState(Cell cell, Cell[,] universeGrid, int universeDimensions)
         {
             var neighbour = new Neighbour(cell, universeDimensions);
             var rows = new List<int> {neighbour.PreviousRow, Constants.FirstRowOrColumn};
-            var columns = new List<int> {neighbour.PreviousColumn, neighbour.NextColumn};
+            var columns = new List<int> {Constants.FirstRowOrColumn, neighbour.PreviousColumn};
             var neighbourCellsState = neighbour.GetNeighbourCellsState(universeGrid, rows, columns);
             return neighbourCellsState;
         }
