+++ ConwaysGameOfLife/Controller/Orientations/Strategies/TopRightCornerStrategy.cs
@@ -1,14 +1,15 @@
 using System.Collections.Generic;
+using ConwaysGameOfLife.Model;
 
-namespace ConwaysGameOfLife.Orientations
+namespace ConwaysGameOfLife.Controller.Orientations.Strategies
 {
-    public static class MiddleStrategy
+    public static class TopRightCornerStrategy
     {
         public static List<State> GetNeighbourCellsState(Cell cell, Cell[,] universeGrid, int universeDimensions)
         {
             var neighbour = new Neighbour(cell, universeDimensions);
-            var rows = new List<int> {neighbour.PreviousRow, neighbour.NextRow};
-            var columns = new List<int> {neighbour.PreviousColumn, neighbour.NextColumn};
+            var rows = new List<int> {neighbour.LastRowOrColumn, neighbour.NextRow};
+            var columns = new List<int> {Constants.FirstRowOrColumn, neighbour.PreviousColumn};
             var neighbourCellsState = neighbour.GetNeighbourCellsState(universeGrid, rows, columns);
             return neighbourCellsState;
         }
