+++ ConwaysGameOfLife/Model/Seed.cs
@@ -1,4 +1,4 @@
-namespace ConwaysGameOfLife
+namespace ConwaysGameOfLife.Model
 {
     public class Seed
     {
