+++ ConwaysGameOfLife/Model/State.cs
@@ -1,4 +1,4 @@
-namespace ConwaysGameOfLife
+namespace ConwaysGameOfLife.Model
 {
     public enum State
     {
