+++ ConwaysGameOfLife/View/GameConsole.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace ConwaysGameOfLife
+namespace ConwaysGameOfLife.View
 {
     public class GameConsole : IGameConsole
     {
