+++ ConwaysGameOfLifeTest/CellTest.cs
@@ -1,5 +1,5 @@
 using System;
-using ConwaysGameOfLife;
+using ConwaysGameOfLife.Model;
 using Xunit;
 
 namespace ConwaysGameOfLifeTest
