+++ ConwaysGameOfLifeTest/NeighbourTest.cs
@@ -1,5 +1,7 @@
 using System;
-using ConwaysGameOfLife;
+using ConwaysGameOfLife.Controller;
+using ConwaysGameOfLife.Model;
+using ConwaysGameOfLife.View;
 using Moq;
 using Xunit;
