+++ ConwaysGameOfLifeTest/SeedCreatorTest.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Collections.Generic;
-using ConwaysGameOfLife;
+using ConwaysGameOfLife.Controller;
+using ConwaysGameOfLife.Model;
+using ConwaysGameOfLife.View;
 using Moq;
 using Xunit;
